Update Q1.Paying User Retention.sql
This SQL code generates a report that calculates the number of paying users retained for each day. It first identifies the date of the first visit for each user and calculates whether each subsequent visit by a user is a repeat visit or a new visit. Finally, it groups the data by event date and counts the number of unique user IDs who made a purchase on a given day and also made a purchase on a previous day (i.e., are retained).