Code Description
This SQL code generates a report that calculates the income earned from in-app purchases (IAPs) for each day. It does this by first creating a view that assigns a price to each type of IAP. The code then sums the prices of all IAPs for each day and orders the results by date.