Code Description
This SQL code generates a report that breaks down the income earned from in-app purchases (IAPs) by age group and gender. It does this by first defining age groups based on the age of the user. The code then joins data from the event, Account, and Product tables to identify the price of each IAP and the age and gender of the user who made the purchase. Finally, the data is grouped by age group and gender and the total income for each group is summed and ordered by age group and gender.